Create action, attachment and test model factories
Action factories includes actions, keywords and reports models that can be safely created separately.

Attachments factories are DataAttachment (based on binary data) and LinkAttachment (based on web links).

Test Model factory is created as well because not only is necessary to link a resource to the recently
created attachment factory but also is very usefull with test when non specific model is necessary
(typically a M2O reference-like field).