[BREAKING] Do not hook workflow transitions in _after_postgeneration
This formula blocks the possibility to create PostDeclarations
interleaved with state transitions. Namely: sale payments ought to be
created after a sale is quoted but before it is confirmed/processed.

Split the state transitions in different inherited factories. This is a
breaking change as the default behaviour exhibited by the Sale and
Purchase factories is now featured by the SaleDraft and PurchaseDraft
factories instead.